+++ src/BuildingRegistry.Projector/Infrastructure/Startup.cs
@@ -16,6 +16,7 @@ namespace BuildingRegistry.Projector.Infrastructure
     using Microsoft.Extensions.DependencyInjection;
     using Microsoft.Extensions.Diagnostics.HealthChecks;
     using Microsoft.Extensions.Logging;
+    using Microsoft.Extensions.Hosting;
     using Modules;
     using Swashbuckle.AspNetCore.Swagger;
     using System;
@@ -119,8 +120,8 @@ namespace BuildingRegistry.Projector.Infrastructure
         public void Configure(
             IServiceProvider serviceProvider,
             IApplicationBuilder app,
-            IHostingEnvironment env,
-            IApplicationLifetime appLifetime,
+            IWebHostEnvironment env,
+            IHostApplicationLifetime appLifetime,
             ILoggerFactory loggerFactory,
             IApiVersionDescriptionProvider apiVersionProvider,
             ApiDataDogToggle datadogToggle,